requirements.txt
# Core AI/ML Libraries
torch>=2.0.0
torchvision>=0.15.0

# GOOGLE GEMINI FLASH - Ultra-Fast, Free LLM (1,500 requests/day)
# Perfect for startups: Free tier, fine-tunable, commercially licensed
google-generativeai>=0.3.0

# Multi-Language Support
langdetect>=1.0.9
googletrans==4.0.0rc1

# 3D Avatar Visualization
pygame>=2.5.0
websockets>=12.0

# Speech Recognition & TTS
SpeechRecognition>=3.10.0
edge-tts>=6.1.0
sounddevice>=0.4.6
soundfile>=0.12.1
PyAudio>=0.2.13
whisperx
pyannote.audio

# Application Control
AppOpener>=1.5
pyautogui>=0.9.54

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
requests-toolbelt>=1.0.0
emoji>=2.8.0

# Optional - Install only if needed
# tuya-connector-python  # For smart home control
# vectordb  # For advanced memory (fallback included in code)
